refactor: renamed app
