+++ wallets/rn_cli_wallet/android/app/src/debug/java/com/walletconnect/web3wallet/rnsample/ReactNativeFlipper.java
@@ -4,7 +4,7 @@
  * <p>This source code is licensed under the MIT license found in the LICENSE file in the root
  * directory of this source tree.
  */
-package com.rn_cli_wallet;
+package com.walletconnect.web3wallet.rnsample;
 
 import android.content.Context;
 import com.facebook.flipper.android.AndroidFlipperClient;
