+++ wallets/rn_cli_wallet/android/app/src/main/java/com/walletconnect/web3wallet/rnsample/MainApplication.java
@@ -1,4 +1,4 @@
-package com.rn_cli_wallet;
+package com.walletconnect.web3wallet.rnsample;
 
 import android.app.Application;
 import com.facebook.react.PackageList;
