+++ wallets/rn_cli_wallet/android/app/src/main/res/values/strings.xml
@@ -1,3 +1,3 @@
 <resources>
-    <string name="app_name">rn_cli_wallet</string>
+    <string name="app_name">RN Web3Wallet</string>
 </resources>
