+++ wallets/rn_cli_wallet/app.json
@@ -1,4 +1,4 @@
 {
-  "name": "rn_cli_wallet",
-  "displayName": "rn_cli_wallet"
+  "name": "RN Web3Wallet",
+  "displayName": "RN Web3Wallet"
 }
