+++ wallets/rn_cli_wallet/ios/RNWeb3Wallet.xcworkspace/contents.xcworkspacedata
@@ -2,7 +2,7 @@
 <Workspace
    version = "1.0">
    <FileRef
-      location = "group:rn_cli_wallet.xcodeproj">
+      location = "group:RNWeb3Wallet.xcodeproj">
    </FileRef>
    <FileRef
       location = "group:Pods/Pods.xcodeproj">
