+++ wallets/rn_cli_wallet/ios/RNWeb3Wallet/AppDelegate.mm
@@ -6,7 +6,7 @@
 
 - (BOOL)application:(UIApplication *)application didFinishLaunchingWithOptions:(NSDictionary *)launchOptions
 {
-  self.moduleName = @"rn_cli_wallet";
+  self.moduleName = @"RN Web3Wallet";
   // You can add your custom initial props in the dictionary below.
   // They will be passed down to the ViewController used by React Native.
   self.initialProps = @{};
