+++ wallets/rn_cli_wallet/ios/RNWeb3Wallet/Images.xcassets/AppIcon.appiconset/iOS App Icon_20pt@3x.png
