+++ wallets/rn_cli_wallet/ios/RNWeb3Wallet/Images.xcassets/AppIcon.appiconset/iOS App Icon_60pt@2x.png
