+++ wallets/rn_cli_wallet/ios/RNWeb3Wallet/Images.xcassets/Contents.json
