+++ wallets/rn_cli_wallet/ios/RNWeb3WalletTests/Info.plist
