+++ wallets/rn_cli_wallet/package.json
@@ -1,5 +1,5 @@
 {
-  "name": "rn_cli_wallet",
+  "name": "rnweb3wallet",
   "version": "0.0.2",
   "private": true,
   "scripts": {
